+++ daxpay-multi/daxpay-multi-service/src/main/java/cn/daxpay/multi/service/dao/order/transfer/TransferOrderMapper.java
@@ -9,6 +9,8 @@ import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Select;
 
+import java.math.BigDecimal;
+
 /**
  *
  * @author xxm
@@ -21,5 +23,5 @@ public interface TransferOrderMapper extends MPJBaseMapper<TransferOrder> {
      * 查询转账总金额
      */
     @Select("select sum(amount) from pay_transfer_order ${ew.customSqlSegment}")
-    Integer getTotalAmount(@Param(Constants.WRAPPER) QueryWrapper<TransferOrderQuery> generator);
+    BigDecimal getTotalAmount(@Param(Constants.WRAPPER) QueryWrapper<TransferOrderQuery> generator);
 }
