+++ daxpay-multi/daxpay-multi-service/src/main/java/cn/daxpay/multi/service/service/order/transfer/TransferOrderQueryService.java
@@ -19,6 +19,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 import java.util.Optional;
 
@@ -100,7 +101,7 @@ public class TransferOrderQueryService {
     /**
      * 查询支付总金额
      */
-    public Integer getTotalAmount(TransferOrderQuery param) {
+    public BigDecimal getTotalAmount(TransferOrderQuery param) {
         return transferOrderManager.getTotalAmount(param);
     }
 }
